requirements.txt
streamlit
google-genai
pydantic
python-dotenv
schedule